% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_functions.R
\name{add_marker_genes_into_compendium}
\alias{add_marker_genes_into_compendium}
\title{add_marker_genes_into_compendium}
\usage{
add_marker_genes_into_compendium(marker.gene.compendium, marker.genes,
  species = "hsapiens", data.format = "vector",
  geneID = "ensembl_gene_id")
}
\arguments{
\item{marker.gene.compendium}{A list consisting of marker gene compendium, species, compendium common genes, as returned by make_marker_gene_compendium.}

\item{marker.genes}{A list or vector containing marker genes. For a list, each element denotes cell type or tissue and contains a vector of marker gene names. For a vector, it is the names of the marker genes of a single cell or tissue. The name of this cell type or tissue will need to be manually updated.}

\item{species}{The species abbreviation name of the marker genes.}

\item{data.format}{The data format of the marker.genes, either 'vector' or 'list'. Default is "vector"}

\item{geneID}{The code for the type of gene IDs used by sample.data, as used by the biomaRt database. To find the valid codes for gene IDs for a species, please see the find_valid_geneID() function of the C3 package. Default is "ensembl_gene_id".}
}
\value{
This function returns a list by adding the new sample data into the marker gene compendium ('marker.genes.matrix').
}
\description{
This function adds a set of marker genes to a marker gene compendium. The marker gene data can be a matrix with more than one cell type / tissue or a vector for a single cell type / tissue.
}
\details{
This function adds a set of marker genes to a marker gene compendium. The marker gene data can be a list with more than one cell type / tissue or a vector for a single cell type / tissue. This function adds only those marker genes that are exist in the marker gene compendium.
}
\examples{
## Here we will use "human.encode.data" and "human.encode.data.descriptor" from C3 repository to make the compendium.
## These data sets are loaded automatically with the package.
## The gene expression data sets are in list format. Here every list element contains transcriptomic profile data of a cell or tissue type.
human.compendium<-make_gene_expression_compendium(gene.expression.data=human.encode.data, experiment.descriptor=human.encode.data.descriptor, expression.data.format="list")

## Next we will make the marker gene compendium from the human.compendium
human.marker.gene.compendium<-make_marker_gene_compendium(compendium.data=human.compendium, specific.cutoff = 0.05, top.expressed.genes = 500)

## Finally we will add the marker genes of a cell type into the compendium
marker.genes<-c("CRYAA", "CRYAB", "CRYBB3", "PAX6", "SOX2", "PROX1", "CRYBA1", "BFSP2", "CRYBB2", "BIRC6")
human.marker.gene.compendium.update<-add_marker_genes_into_compendium(marker.gene.compendium = human.marker.gene.compendium, marker.genes = marker.genes, data.format = "vector", geneID = "external_gene_name")

## Here, we will see that all marker genes are added as a 'newly added' cell type in the marker gene compendium except 'PROX1' as it does not exist in the universe of the marker gene compendium
names(which(human.marker.gene.compendium.update$marker.genes.matrix[,1]==1))

}
