% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{c3_bar}
\alias{c3_bar}
\title{Bar Plot}
\usage{
c3_bar(c3, stacked = FALSE, rotated = FALSE, bar_width = 0.6, zerobased = TRUE)
}
\arguments{
\item{c3}{c3 htmlwidget object}

\item{stacked}{boolean place bars on top of each other}

\item{rotated}{boolean use to make x-axis vertical}

\item{bar_width}{numeric pixel width of bars}

\item{zerobased}{boolean}
}
\value{
c3
}
\description{
Add bars to a C3 plot
}
\examples{
data.frame(a=c(1,2,3,2),b=c(2,3,1,5)) \%>\%
  c3() \%>\%
  c3_bar(stacked = TRUE)

}
