% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{c3_chart_size}
\alias{c3_chart_size}
\title{Chart Size}
\usage{
c3_chart_size(
  c3,
  left = NULL,
  right = NULL,
  top = NULL,
  bottom = NULL,
  width = NULL,
  height = NULL,
  ...
)
}
\arguments{
\item{c3}{c3 htmlwidget object}

\item{left}{integer padding pixels}

\item{right}{integer padding pixels}

\item{top}{integer padding pixels}

\item{bottom}{integer padding pixels}

\item{width}{integer pixels}

\item{height}{integer pixels}

\item{...}{additional options passed to the padding and size objects}
}
\value{
c3
}
\description{
Modify the size of the chart within the htmlwidget area. Generally charts size
to the div in which they are placed. These options enable finer scale sizing with the div
}
\examples{
data.frame(a = c(1,2,3,2), b = c(2,4,1,5)) \%>\%
  c3() \%>\%
  c3_chart_size(width = 600, height = 200)

}
