% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{c3_color}
\alias{c3_color}
\title{Color Palette}
\usage{
c3_color(c3, colors)
}
\arguments{
\item{c3}{c3 htmlwidget object}

\item{colors}{character vector of colors}
}
\value{
c3
}
\description{
Manually assign colors
}
\examples{
data.frame(a = c(1,2,3,2), b = c(2,4,1,5)) \%>\%
  c3() \%>\%
  c3_color(c('red','black'))

}
