% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{c3_donut}
\alias{c3_donut}
\title{Donut Charts}
\usage{
c3_donut(
  c3,
  expand = TRUE,
  title = NULL,
  width = NULL,
  show = TRUE,
  threshold = NULL,
  format = NULL,
  ...
)
}
\arguments{
\item{c3}{c3 htmlwidget object}

\item{expand}{boolean expand segment on hover}

\item{title}{character}

\item{width}{integer pixels width of donut}

\item{show}{boolean show labels}

\item{threshold}{numeric proportion of segment to hide label}

\item{format}{character label js function, wrap character or character vector in JS()}

\item{...}{additional values passed to the donut label object}
}
\value{
c3
}
\description{
Create simple Donut charts
}
\examples{
data.frame(red=20,green=45,blue=10) \%>\%
  c3() \%>\%
  c3_donut(title = 'Colors')

}
