% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_functions.R
\docType{data}
\name{c3_params}
\alias{c3_params}
\alias{biomart.ID}
\title{c3_params}
\format{An object of class \code{character} of length 1.}
\usage{
biomart.ID
}
\description{
When Ensembl move servers (rarely) or it is down (less rare) these variable may need to be changed, for example to an archived version
}
\keyword{datasets}
