% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{c3_pie}
\alias{c3_pie}
\title{Pie Charts}
\usage{
c3_pie(c3, show = TRUE, threshold = NULL, format = NULL, expand = TRUE, ...)
}
\arguments{
\item{c3}{c3 htmlwidget object}

\item{show}{boolean show labels}

\item{threshold}{numeric proportion of segment to hide label}

\item{format}{character label js function, wrap character or character vector in JS()}

\item{expand}{boolean expand segment on hover}

\item{...}{additional values passed to the pie label object}
}
\value{
c3
}
\description{
C3 Pie Charts
}
\examples{
data.frame(red = 20, green = 45, blue = 10) \%>\%
  c3() \%>\%
  c3_pie()

}
