% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{c3_scatter}
\alias{c3_scatter}
\title{Scatter Plots}
\usage{
c3_scatter(c3, ...)
}
\arguments{
\item{...}{}
}
\examples{
\dontrun{
  iris \%>\%
    c3(x = 'Sepal_Length',
       y = 'Sepal_Width',
       group = 'Species') \%>\%
    c3_scatter()
  }

}
