% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{c3_viridis}
\alias{c3_viridis}
\title{Viridis Palette}
\usage{
c3_viridis(c3, pal = "D")
}
\arguments{
\item{c3}{c3 htmlwidget object}

\item{pal}{character palette options}
}
\value{
c3
}
\description{
Use Viridis palette options
}
\examples{
data.frame(a = c(1,2,3,2), b = c(2,4,1,5)) \%>\%
  c3() \%>\%
  c3_viridis()

}
