% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3.R
\name{c3ChartProxy}
\alias{c3ChartProxy}
\title{Create a C3 proxy object}
\usage{
c3ChartProxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{name of the C3 widget}

\item{session}{valid session object}
}
\value{
C3 proxy object
}
\description{
Create a C3 proxy object
}

