% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3Gauge.R
\name{C3Gauge}
\alias{C3Gauge}
\title{C3Gauge widget}
\usage{
C3Gauge(value, width = NULL, height = NULL)
}
\arguments{
\item{value}{value of gauge}

\item{width}{chart width}

\item{height}{chart height}
}
\description{
C3Gauge widget
}

