% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3LineBarChart.R
\name{C3LineBarChart-shiny}
\alias{C3LineBarChart-shiny}
\alias{C3LineBarChartOutput}
\alias{renderC3LineBarChart}
\title{Shiny bindings for C3LineBarChart}
\usage{
C3LineBarChartOutput(outputId, width = "100\%", height = "400px")

renderC3LineBarChart(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a C3LineBarChart}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using C3LineBarChart within Shiny
applications and interactive Rmd documents.
}

