% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3LineBarChart.R
\name{C3LineBarChart}
\alias{C3LineBarChart}
\title{C3LineBarChart widget}
\usage{
C3LineBarChart(dataset, colors, width = NULL, height = NULL)
}
\arguments{
\item{dataset}{dataframe with Time colom (Year-month-day) and one numeric column per time series. Column names are used as keys.}

\item{colors}{named list providing colors e.g. list(Total="gray",GREEN="Green",RED="Red")}

\item{width}{chart width}

\item{height}{chart height}
}
\description{
C3LineBarChart widget
}

