% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3Pie.R
\name{C3Pie}
\alias{C3Pie}
\title{C3Pie widget}
\usage{
C3Pie(values, legendPosition = "bottom", width = NULL, height = NULL)
}
\arguments{
\item{values}{named list with values e.g. list('data1'=30,'data2'=120)}

\item{legendPosition}{either top-left, top-right, bottom-left or bottom-right}

\item{width}{chart width}

\item{height}{chart height}
}
\description{
C3Pie widget
}

