% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{check_stacked}
\alias{check_stacked}
\title{Check groups for stacked plots}
\usage{
check_stacked(c3, stacked)
}
\arguments{
\item{c3}{c3 htmlwidget object}

\item{stacked}{boolean}
}
\value{
c3 object
}
\description{
For plots where stacking is required this function will
define the columns to be stacked based on column headers.
}
