% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_functions.R
\name{convert_compendium_to_gene_sets}
\alias{convert_compendium_to_gene_sets}
\title{convert_compendium_to_gene_sets}
\usage{
convert_compendium_to_gene_sets(specific.compendium)
}
\arguments{
\item{specific.compendium}{A marker gene matrix as returned by 'make_marker_gene_compendium' (i.e., marker.genes.matrix). That is, a matrix containing the presence (1) / absence (0) information for marker genes in each cell type / tissue. Rows denote genes, columns denote individual tissues or cell types.}
}
\value{
This function returns a list  with two sub-list consisting of marker gene gene-sets and the gene universe for each cell type or tissue, with elements of the sub-lists corresponding to columns of the input matrix.
}
\description{
This function converts a marker gene matrix into marker gene gene-sets and gene universe.
}
\details{
This function converts a marker gene matrix into marker gene gene-sets and gene universe.
}
\examples{
specific.compendium<-matrix(sample(0:1, 100, replace=T), 10, 10)
rownames(specific.compendium)<-c("ENSG00000100053", "ENSG00000109846", "ENSG00000244752", "ENSG00000138083", "ENSG00000181449", "ENSG00000160202", "ENSG00000007372", "ENSG00000117707", "ENSG00000101144", "ENSG00000277354")
colnames(specific.compendium)<-c("cell1", "cell2", "cell3", "cell4", "cell5", "cell6", "cell7", "cell8", "cell9", "cell10")
convert_compendium_to_gene_sets(specific.compendium)

}
