% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_functions.R
\name{convert_geneID_to_ensemblID}
\alias{convert_geneID_to_ensemblID}
\title{convert_geneID_to_ensemblID}
\usage{
convert_geneID_to_ensemblID(user.cell.data, species = "hsapiens",
  ensemblId.toMapping = "ensembl_gene_id",
  geneID.forMapping = "external_gene_name", collapse.method = "max",
  forMarkerGenes = F, markerGeneFormat = "vector")
}
\arguments{
\item{user.cell.data}{Either valid gene IDs for marker genes or cell/tissue profile with valid gene IDs}

\item{species}{The species of the given gene IDs or cell/tissue profile, default is "hsapiens"}

\item{ensemblId.toMapping}{The valid gene ID names (that has ensembl id mapping in the biomaRt database) to which the user.cell.data will be converted, default is "ensembl_gene_id". For more details to get the valid gene ids for a species please see the find_valid_geneID() function of the C3 package.}

\item{geneID.forMapping}{The valid gene ID names (that has ensembl id mapping in the biomaRt database) of the user.cell.data, default is "external_gene_name". For more details to get the valid gene ids for a species please see the find_valid_geneID() function of the C3 package.}

\item{collapse.method}{Used when one ensembl_gene_id has more then one probes, usually two options are used, maximum probe value (i.e., "max") or average of probe values (i.e., "mean"), default is "max". Not used for marker gene option.}

\item{forMarkerGenes}{Boolean with either T for marker genes conversion or F for cell/tissue profile symbol conversion, default is F}

\item{markerGeneFormat}{Data format of marker genes either 'list' or 'vector', used only during marker genes conversion, default is "vector"}
}
\value{
This function returns either ensembl gene IDs for the corresponding gene IDs of marker genes or cell/tissue profiles with converted ensembl gene IDs
}
\description{
This function converts either a vector of gene IDs, a matrix of gene or a named cell / tissue expression profile or matrix to ensembl IDs
}
\details{
TThis function converts a vector of gene IDs or a named cell/tissue expression profile to ensembl IDs
}
\examples{
marker.genes<-c("CRYAA", "CRYAB", "CRYBB3", "PAX6", "SOX2", "PROX1", "SIX3", "CRIM1", "CRYBB2", "BMP7")
convert_geneID_to_ensemblID(user.cell.data=marker.genes, species = "hsapiens", forMarkerGenes = T, markerGeneFormat = "vector")

}
