% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_functions.R
\name{find_valid_geneID}
\alias{find_valid_geneID}
\title{find_valid_geneID}
\usage{
find_valid_geneID(species = "hsapiens")
}
\arguments{
\item{species}{Species name in the form "hsapiens"}
}
\value{
This helper function returns a data frame with two columns containing all valid gene IDs and their description that have mapping to Ensembl gene IDs for the given species
}
\description{
This helper function returns a data frame with two columns containing all valid gene IDs and their description that have mapping to Ensembl gene IDs for the given species. For supported species that have ensembl id mapping run 'supportedSpecies' (this command is from XGSA package), it will return all the names of the supported species. For each species there exists separate ensembl datasets.
}
\details{
This helper function returns a data frame with two columns containing all valid gene IDs and their description that have mapping to Ensembl gene IDs for the given species. For supported species that have ensembl id mapping run 'supportedSpecies' (this command is from XGSA package), it will return all the names of the supported species. For each species there exists separate ensembl dataset.
}
\examples{
human.valid.gene.IDs<-find_valid_geneID(species = "hsapiens")
head(human.valid.gene.IDs)
}
