% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hawse.human.lens.data}
\alias{hawse.human.lens.data}
\title{Hawse human lens data}
\format{An object of matrix containing gene expression value of LEC and LFC.}
\usage{
data(hawse.human.lens.data)
}
\description{
Data from the Hawse et al human lens,
contains transcriptomic profiles data of lens epithelial cell (LEC) and lens fiber cell (LFC).
Each cell has 3 biological replicates.
}
\examples{
head(hawse.human.lens.data)
}
\references{
Hawse et al, Mol Vis. 2005 Apr 18; 11: 274–283
(\href{https://www.ncbi.nlm.nih.gov/pubmed/15851978}{PubMed})
}
\keyword{datasets}
