% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hoang.mouse.lens.data}
\alias{hoang.mouse.lens.data}
\title{Hoang mouse lens data}
\format{An object of matrix containing average gene expression value of LEC and LFC.}
\usage{
data(hoang.mouse.lens.data)
}
\description{
Data from the Hoang et al mouse lens,
contains transcriptomic profiles data of lens epithelial cell (LEC) and lens fiber cell (LFC).
Each cell contains average expression value of 3 biological replicates.
}
\examples{
head(hoang.mouse.lens.data)
}
\references{
Hoang et al, Mol Vis. 2014 Nov 4;20:1491-517
(\href{https://www.ncbi.nlm.nih.gov/pubmed/25489224}{PubMed})
}
\keyword{datasets}
