% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human.encode.data.descriptor}
\alias{human.encode.data.descriptor}
\title{Human ENCODE data descriptor}
\format{An object of vector containing 144 experiment names.}
\usage{
data(human.encode.data.descriptor)
}
\description{
Data from the human ENCODE project, contains experiment descriptors of the data sets.
}
\examples{
head(human.encode.data.descriptor)
}
\references{
The ENCODE Project Consortium, Nature 2012 Sep 6;489(7414):57-74
(\href{https://www.ncbi.nlm.nih.gov/pubmed/22955616}{PubMed})
}
\keyword{datasets}
