% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human.encode.data}
\alias{human.encode.data}
\title{Human ENCODE data}
\format{An object of list containing 144 data frames of the gene expression profiles.}
\usage{
data(human.encode.data)
}
\description{
Data from the human ENCODE project, contains transcriptomic profiles of 144 different cell lines and/or tissues.
}
\examples{
summary(human.encode.data)
}
\references{
The ENCODE Project Consortium, Nature 2012 Sep 6;489(7414):57-74
(\href{https://www.ncbi.nlm.nih.gov/pubmed/22955616}{PubMed})
}
\keyword{datasets}
