% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend}
\alias{legend}
\title{Modify plot elements that relate to the legend.}
\usage{
legend(x, ...)
}
\description{
This is an S3 method. The c3 legend is on by default, this function allows the
legend to be removed, or other legend attributes to be set.
}
\seealso{
Other legend: \code{\link{legend.c3}}
}
