% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_functions.R
\name{make_ENSEMBL_symbol_map}
\alias{make_ENSEMBL_symbol_map}
\title{make_ENSEMBL_symbol_map}
\usage{
make_ENSEMBL_symbol_map(species, ensemblId.toMapping, geneID.forMapping)
}
\arguments{
\item{species}{Species name in the form 'hsapiens'}

\item{ensemblId.toMapping}{The valid gene ID names to which the user cell data will be converted, default is "ensembl_gene_id"}

\item{geneID.forMapping}{The valid gene ID names of the user cell data, default is "external_gene_name"}
}
\value{
This helper function returns a data frame with two columns containing all Ensembl gene IDs and their corresponding gene IDs for the given species
}
\description{
This helper function returns a data frame with two columns containing all Ensembl gene IDs and their corresponding gene IDs for the given species
}
\details{
This helper function returns a data frame with two columns containing all Ensembl gene IDs and their corresponding gene IDs for the given species
}
\examples{
human.ensembl.symbol.map<-make_ENSEMBL_symbol_map('hsapiens', ensemblId.toMapping="ensembl_gene_id", geneID.forMapping="external_gene_name")
head(human.ensembl.symbol.map)

}
