% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_functions.R
\name{make_marker_gene_compendium}
\alias{make_marker_gene_compendium}
\title{make_marker_gene_compendium}
\usage{
make_marker_gene_compendium(compendium.data, specific.cutoff = 0.05,
  top.expressed.genes = 500)
}
\arguments{
\item{compendium.data}{A list consisiting of gene expression compendium (matrix) and compendium data species name, as returned by make_gene_expression_compendium.}

\item{specific.cutoff}{A percentage threshold representing the maximum number of cell types in which a gene can be highly expressed in order for it to be considered a marker gene. Default is 0.05.}

\item{top.expressed.genes}{The number of highest expressed genes from each cell/tissue that will be considered for marker gene status. Default is 500.}
}
\value{
This function returns a  list with  cell/tissue marker gene matrix, compendium common genes, species and top number of expressed genes. In the marker gene compendium (marker.genes.matrix), rows denote genes and columns denote cell types / tissues, containing presence (1) or absence (0) of marker genes for each cell type / tissue.
}
\description{
This function transforms a gene expression compendium into a marker gene compendium, using certain threshold parameters.
}
\details{
This function transforms a gene expression compendium into a marker gene compendium, using certain threshold parameters. The two parameters decide what is considered 'highly expressed' and how few cell types is considered 'unique' enough to be a marker gene.
}
\examples{
## First we will load the human ENCODE data set and their descriptor from C3 repository
load("https://github.com/VCCRI/C3/data/all_FPKMs_human_cleanID.RData")
load("https://github.com/VCCRI/C3/data/experiment.descriptors.humanEncode.RData")

## Then we will make the compendium using these data sets
## The gene expression data sets are in list format. Here every list element denotes a cell or tissue type
human.compendium<-make_gene_expression_compendium(gene.expression.data=all_FPKMs_clean_ID, experiment.descriptor=experiment.descriptors.humanEncode, expression.data.format="list")

## Finally we will make the marker gene compendium from the human.compendium
human.marker.gene.compendium<-make_marker_gene_compendium(compendium.data=human.compendium, specific.cutoff = 0.05, top.expressed.genes = 500)
summary(human.marker.gene.compendium)

}

