% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{a \code{\link{c3}} object}

\item{rhs}{a pie settings function}
}
\description{
Imports the pipe operator from magrittr.
}
\examples{
data.frame(a=c(1,2,3,2),b=c(2,3,1,5)) \%>\%
  c3()
}
