% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{RColorBrewer}
\alias{RColorBrewer}
\alias{RColorBrewer.c3}
\title{RColorBrewer Palette}
\usage{
RColorBrewer(c3, pal = "Spectral")

\method{RColorBrewer}{c3}(c3, pal = "Spectral")
}
\arguments{
\item{c3}{c3 htmlwidget object}

\item{pal}{character palette must match `RColorBrewer::brewer.pal.info`}
}
\value{
c3
}
\description{
Use RColorBrewer palettes
}
\examples{
data.frame(a = c(1,2,3,2), b = c(2,4,1,5), c = c(5,3,4,1)) \%>\%
  c3() \%>\%
  RColorBrewer()

}
\seealso{
Other c3: 
\code{\link{c3}()},
\code{\link{grid}()},
\code{\link{legend}()},
\code{\link{region}()},
\code{\link{subchart}()},
\code{\link{tooltip}()},
\code{\link{xAxis}()},
\code{\link{zoom}()}
}
\concept{RColorBrewer}
\concept{c3}
