% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{RColorBrewer}
\alias{RColorBrewer}
\title{RColorBrewer Palette}
\usage{
RColorBrewer(x, ...)
}
\description{
Use RColorBrewer palettes. This is an S3 method.
}
\seealso{
Other RColorBrewer: \code{\link{RColorBrewer.c3}}
}
