% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_functions.R
\name{select_query_sample_specific_genes}
\alias{select_query_sample_specific_genes}
\title{select_query_sample_specific_genes}
\usage{
select_query_sample_specific_genes(processed.queryData, comSpecies, topExpGenes,
  compendiumCommonGenes)
}
\arguments{
\item{processed.queryData}{A list containing the average gene expression values for the cell types and/or tissues with ensembl gene IDs and the species name, as returned by 'preprocess_querydata'.}

\item{comSpecies}{The species of the compendium which the query is being compared to. Default is "hsapiens".}

\item{topExpGenes}{The number of highest expressed genes from each cell/tissue that will be considered for marker gene status.}

\item{compendiumCommonGenes}{The list of commonly expressed (non-marker) genes as determined from the compendium, which will be removed from the list of highly expressed sample genes.}
}
\value{
This function returns a list containing XGSA data sets ready to test with the compendium data.
}
\description{
This function converts the query data into a list of XGSA data sets containing the marker genes for each sample.
}
\details{
This function converts the query data into a list of XGSA data sets containing the marker genes for each sample.
}
\examples{
Used within C3 framework

}

