% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3StackedChart.R
\name{setColors}
\alias{setColors}
\title{Sets the colors of an existing C3 chart}
\usage{
setColors(chartProxy, colors)
}
\arguments{
\item{chartProxy}{C3 chart proxy object}

\item{colors}{named list with colors for the data series in the chart}
}
\value{
C3 proxy object
}
\description{
Sets the colors of an existing C3 chart
}

