% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3StackedChart.R
\name{setGroups}
\alias{setGroups}
\title{Set chart groups}
\usage{
setGroups(chartProxy, groups)
}
\arguments{
\item{chartProxy}{C3 chart proxy object}

\item{groups}{vector containing groups used to stack the chart}
}
\value{
C3 proxy object
}
\description{
Set chart groups
}

