% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3StackedChart.R
\name{setTypes}
\alias{setTypes}
\title{Set chart types}
\usage{
setTypes(chartProxy, info)
}
\arguments{
\item{chartProxy}{C3 chart proxy object}

\item{message}{list containing data series types}
}
\value{
C3 proxy object
}
\description{
Set chart types
}

