% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subchart.R
\name{subchart}
\alias{subchart}
\title{Add subchart}
\usage{
subchart(x, ...)
}
\description{
Subcharts are defined in multiple axis by passing a single `data.frame`. Subcharts are listed as an
experimental feature in the \href{http://c3js.org/reference.html#subchart-onbrush}{C3 documentation}). This is an S3 method.
}
\seealso{
Other subchart: \code{\link{subchart.c3}}
}
