% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip}
\alias{tooltip}
\title{C3 Tooltips}
\usage{
tooltip(x, ...)
}
\description{
Modify plot elements that relate to tooltips. This is an S3 method.
}
\seealso{
Other tooltip: \code{\link{tooltip.c3}}
}
