% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3StackedChart.R
\name{transformTo}
\alias{transformTo}
\title{Set all chart types simulaneously}
\usage{
transformTo(chartProxy, type)
}
\arguments{
\item{chartProxy}{C3 chart proxy object}

\item{message}{list containing data series type}
}
\value{
C3 proxy object
}
\description{
Set all chart types simulaneously
}

