% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{y2Axis}
\alias{y2Axis}
\title{Modify plot elements that relate to the second y-axis. S3 Method}
\usage{
y2Axis(x, ...)
}
\description{
This is an S3 method.
}
\seealso{
Other axis: \code{\link{xAxis.c3}}, \code{\link{yAxis}}
}
