% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{yAxis}
\alias{yAxis}
\title{C3 yaxis}
\usage{
yAxis(x, ...)
}
\description{
Modify plot elements that relate to the y-axis. This is an S3 method.
}
\seealso{
Other axis: \code{\link{xAxis.c3}}, \code{\link{y2Axis}}
}
