% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3StackedChart.R
\name{zoomChart}
\alias{zoomChart}
\title{Zooms an existing C3Chart from shiny}
\usage{
zoomChart(chartProxy, minX, maxX)
}
\arguments{
\item{minX}{date start}

\item{maxX}{date end}
}
\value{
C3 proxy object
}
\description{
Zooms an existing C3Chart from shiny
}

