% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3.R
\name{c3}
\alias{c3}
\title{<Add Title>}
\usage{
c3(data, x = NULL, y = NULL, group = NULL, message = NULL,
  width = 800, height = 300, ...)
}
\arguments{
\item{labels}{character or list with otpions:
 \itemize{
 \item{format}{: list format functions for each parameter label (see \href{http://c3js.org/reference.html#data-labels}{c3 data-labels})}
}}

\item{hide}{booleen or character vector of parameters to hide}
}
\description{
<Add Description>
}
\seealso{
Other c3: \code{\link{RColorBrewer.c3}},
  \code{\link{grid.c3}}, \code{\link{legend.c3}},
  \code{\link{region.c3}}, \code{\link{subchart.c3}},
  \code{\link{tooltip.c3}}, \code{\link{xAxis.c3}},
  \code{\link{zoom.c3}}
}

