% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3.R
\name{c3}
\alias{c3}
\title{C3}
\usage{
c3(data, x = NULL, y = NULL, group = NULL, message = NULL,
  width = NULL, height = NULL, ...)
}
\arguments{
\item{labels}{character or list with otpions:
 \itemize{
 \item{format}{: list format functions for each parameter label (see \href{http://c3js.org/reference.html#data-labels}{c3 data-labels})}
}}

\item{hide}{booleen or character vector of parameters to hide}

\item{onclick}{character js function, wrap character or character vector in JS()}

\item{onmouseover}{character js function, wrap character or character vector in JS()}

\item{onmouseout}{character js function, wrap character or character vector in JS()}
}
\description{
The `c3` package is a wrapper, or \href{http://www.htmlwidgets.org/}{htmlwidget}, for the \href{http://c3js.org/}{c3} javascript charting library by \href{https://github.com/masayuki0812}{Masayuki Tanaka}.
}
\examples{
\dontrun{
data <- data.frame(a=c(1,2,3,2),b=c(2,3,1,5))

data \%>\%
  c3(onclick = htmlwidgets::JS("function(d, element){console.log(d)}"))

data.frame(sugar=20,fat=45,salt=10) \%>\%
  c3(onclick = htmlwidgets::JS("function(d, element){dp = d}")) \%>\%
  c3_pie()
  }

}
\seealso{
Other c3: \code{\link{RColorBrewer.c3}},
  \code{\link{grid.c3}}, \code{\link{legend.c3}},
  \code{\link{region.c3}}, \code{\link{subchart.c3}},
  \code{\link{tooltip.c3}}, \code{\link{xAxis.c3}},
  \code{\link{zoom.c3}}
}
