% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{c3_bar}
\alias{c3_bar}
\title{Bar Plot}
\usage{
c3_bar(c3, stacked = FALSE, rotated = FALSE, bar_width = 0.6,
  zerobased = TRUE)
}
\value{
c3
}
\description{
Bar Plot
}
\examples{
\dontrun{
data.frame(a=c(1,2,3,2),b=c(2,3,1,5)) \%>\%
  c3() \%>\%
  c3_bar()
  }

}

