% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{c3_chart_size}
\alias{c3_chart_size}
\title{Chart Size}
\usage{
c3_chart_size(c3, ...)
}
\arguments{
\item{left}{integer padding pixels}

\item{right}{integer padding pixels}

\item{top}{integer padding pixels}

\item{bottom}{integer padding pixels}

\item{width}{integer pixels}

\item{height}{integer pixels}
}
\value{
c3
}
\description{
Modify the size of the chart within the htmlwidget area. Generally charts size
to the div in which they are placed. These options enable finer scale sizing with the div
}

