% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{c3_donut}
\alias{c3_donut}
\title{Donut Charts}
\usage{
c3_donut(c3, title = NULL, ...)
}
\arguments{
\item{title}{chasracter}

\item{...}{}
}
\description{
Donut Charts
}
\examples{
\dontrun{
data.frame(red=20,green=45,blue=10) \%>\%
  c3() \%>\%
  c3_donut(title = 'Colors')
  }

}

