% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{c3_gauge}
\alias{c3_gauge}
\title{Gauge Charts}
\usage{
c3_gauge(c3, ...)
}
\arguments{
\item{...}{}

\item{label}{list with options:
\itemize{
 \item{show}{: boolean}
 \item{format}{: function, wrap in JS() }
}}

\item{min}{numeric}

\item{max}{numeric}

\item{units}{character appended to numeric value}

\item{width}{integer pixel width of the arc}

\item{pattern}{character vector or pallete of colors}

\item{threshold}{list with options:
\itemize{
 \item{unit}{: character one of 'percent', 'value'}
 \item{max}{: numeric}
 \item{values}{: numeric vector of threhold values for color change}
}}
}
\examples{
\dontrun{
data.frame(red=20,green=45,blue=10) \%>\%
  c3() \%>\%
  c3_donut(title = 'Colors')
  }

}
