% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{c3_mixedGeom}
\alias{c3_mixedGeom}
\title{Title}
\usage{
c3_mixedGeom(c3, types, type = "line", stacked = NULL, ...)
}
\arguments{
\item{types}{list containing key value pairs of column header and plot type}

\item{stacked}{character vector of column headers to stack}

\item{...}{}
}
\description{
Title
}

