% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{c3_scatter}
\alias{c3_scatter}
\title{Scatter Plots}
\usage{
c3_scatter(c3)
}
\arguments{
\item{c3}{c3 htmlwidget object}
}
\value{
c3
}
\description{
For scatter plots options are defined in the `c3` function. Options are limited to x, y and groups
}
\examples{
  iris \%>\%
    c3(x = 'Sepal_Length',
       y = 'Sepal_Width',
       group = 'Species') \%>\%
    c3_scatter()

}
