% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{c3_selection}
\alias{c3_selection}
\title{Data Select}
\usage{
c3_selection(c3, ...)
}
\arguments{
\item{enabled}{boolean}

\item{groued}{boolean}

\item{multiple}{boolean}

\item{draggable}{boolean}

\item{isselectable}{character js function, wrap character or character vector in JS()}
}
\value{
c3
}
\description{
Define options for selecting data within the plot area
}

