% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{c3_selection}
\alias{c3_selection}
\title{Data Select}
\usage{
c3_selection(c3, enabled = FALSE, grouped = FALSE, multiple = FALSE,
  draggable = FALSE, isselectable = JS("function () { return true; }"),
  ...)
}
\arguments{
\item{c3}{c3 htmlwidget object}

\item{enabled}{boolean}

\item{grouped}{boolean}

\item{multiple}{boolean}

\item{draggable}{boolean}

\item{isselectable}{character js function, wrap character or character vector in JS()}

\item{...}{additional options passed to data selection object}
}
\value{
c3
}
\description{
Define options for selecting data within the plot area
}
\examples{
data.frame(a = c(1,2,3,2), b = c(2,3,1,5)) \%>\%
  c3() \%>\%
  c3_selection(enabled = TRUE,
               multiple = TRUE)

}
