% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid.c3}
\alias{grid.c3}
\title{Title}
\usage{
\method{grid}{c3}(c3, axis, ...)
}
\arguments{
\item{axis}{character 'x' or 'y'}

\item{show}{boolean}

\item{lines}{vector of lists each with options:
\itemize{
 \item{value}{: numeric, character or date depending on axis}
 \item{text}{: character (optional)}
 \item{class}{: character css class (optional)}
 \item{position}{: character one of 'start', 'middle', 'end' (optional)}
}}
}
\value{
c3
}
\description{
Title
}
\seealso{
Other c3: \code{\link{c3}}, \code{\link{legend.c3}},
  \code{\link{tooltip.c3}}, \code{\link{xAxis.c3}}

Other grid: \code{\link{grid}}
}

