% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid.c3}
\alias{grid.c3}
\title{Title}
\usage{
\method{grid}{c3}(c3, axis, ...)
}
\arguments{
\item{axis}{character 'x' or 'y'}

\item{show}{boolean}

\item{lines}{dataframe with options:
\itemize{
 \item{value}{: numeric, character or date depending on axis}
 \item{text}{: character (optional)}
 \item{class}{: character css class (optional)}
 \item{position}{: character one of 'start', 'middle', 'end' (optional)}
}}
}
\value{
c3
}
\description{
Title
}
\examples{
\dontrun{
iris \%>\%
 c3(x='Sepal_Length', y='Sepal_Width', group = 'Species') \%>\%
 c3_scatter() \%>\%
 grid('y') \%>\%
 grid('x', show=F, lines = data.frame(value=c(5,6),
                                      text= c('Line 1','Line 2')))
}

}
\seealso{
Other c3: \code{\link{RColorBrewer.c3}}, \code{\link{c3}},
  \code{\link{legend.c3}}, \code{\link{region.c3}},
  \code{\link{subchart.c3}}, \code{\link{tooltip.c3}},
  \code{\link{xAxis.c3}}, \code{\link{zoom.c3}}

Other grid: \code{\link{grid}}
}
