% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend.c3}
\alias{legend.c3}
\title{C3 Legend Options}
\usage{
\method{legend}{c3}(c3, ...)
}
\arguments{
\item{c3}{c3 htmlwidget object}

\item{...}{all variables passed via ...}

\item{hide}{boolean or character of parameters to hide}

\item{position}{character one of 'bottom', 'right', 'inset'}

\item{inset}{list with options:
\itemize{
 \item{anchor}{: character one of 'top-left', 'top-right', 'bottom-left', 'bottom-right'}
 \item{x}{: integer pixels}
 \item{y}{: integer pixels}
 \item{step}{: numeric}
}}

\item{item}{list with options:
\itemize{
 \item{onclick}{: character js function, wrap character or character vector in JS()}
 \item{onmouseover}{: character js function, wrap character or character vector in JS()}
 \item{onmouseout}{: character js function, wrap character or character vector in JS()}
}}
}
\value{
c3
}
\description{
Modify plot elements that relate to the legend. The c3 legend is on by default, this function allows the
legend to be removed, or other legend attributes to be set.
}
\examples{
\dontrun{
iris \%>\%
 c3(x='Sepal_Length', y='Sepal_Width', group = 'Species') \%>\%
 c3_scatter() \%>\%
 legend(position = 'right')
}

}
\seealso{
Other c3: \code{\link{RColorBrewer.c3}}, \code{\link{c3}},
  \code{\link{grid.c3}}, \code{\link{region.c3}},
  \code{\link{subchart.c3}}, \code{\link{tooltip.c3}},
  \code{\link{xAxis.c3}}, \code{\link{zoom.c3}}

Other legend: \code{\link{legend}}
}
