% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend.c3}
\alias{legend.c3}
\title{Title}
\usage{
\method{legend}{c3}(c3, ...)
}
\arguments{
\item{...}{}

\item{hide}{boolean or character of parameters to hide}

\item{position}{character one of 'bottom', 'right', 'inset'}

\item{inset}{list with options:
\itemize{
 \item{anchor}{: character one of 'top-left', 'top-right', 'bottom-left', 'bottom-right'}
 \item{x}{: integer pixels}
 \item{y}{: integer pixels}
 \item{step}{: numeric}
}}

\item{item}{list with options:
\itemize{
 \item{onclick}{: character js function, wrap character or character vector in JS()}
 \item{onmouseover}{: character js function, wrap character or character vector in JS()}
 \item{onmouseout}{: character js function, wrap character or character vector in JS()}
}}
}
\value{
c3
}
\description{
Title
}
\seealso{
Other c3: \code{\link{RColorBrewer.c3}}, \code{\link{c3}},
  \code{\link{grid.c3}}, \code{\link{region.c3}},
  \code{\link{tooltip.c3}}, \code{\link{xAxis.c3}}

Other legend: \code{\link{legend}}
}

