% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend}
\alias{legend}
\title{C3 Legend Options}
\usage{
legend(x, ...)
}
\description{
Modify plot elements that relate to the legend. The c3 legend is on by default, this function allows the
legend to be removed, or other legend attributes to be set. This is an S3 method.
}
\seealso{
Other legend: \code{\link{legend.c3}}
}
