% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_functions.R
\name{make_marker_gene_compendium}
\alias{make_marker_gene_compendium}
\title{make_marker_gene_compendium}
\usage{
make_marker_gene_compendium(compendium.data, specific.cutoff = 0.05,
  top.expressed.genes = 500)
}
\arguments{
\item{compendium.data}{A list consisiting of gene expression compendium (matrix) and compendium data species name, as returned by make_gene_expression_compendium.}

\item{specific.cutoff}{A percentage threshold representing the maximum number of cell types in which a gene can be highly expressed in order for it to be considered a marker gene. Default is 0.05.}

\item{top.expressed.genes}{The number of highest expressed genes from each cell/tissue that will be considered for marker gene status. Default is 500.}
}
\value{
This function returns a  list with  cell/tissue marker gene matrix, compendium common genes, species and top number of expressed genes. In the marker gene compendium (marker.genes.matrix), rows denote genes and columns denote cell types / tissues, containing presence (1) or absence (0) of marker genes for each cell type / tissue.
}
\description{
This function transforms a gene expression compendium into a marker gene compendium, using certain threshold parameters.
}
\details{
This function transforms a gene expression compendium into a marker gene compendium, using certain threshold parameters. The two parameters decide what is considered 'highly expressed' and how few cell types is considered 'unique' enough to be a marker gene.
}
\examples{
## Here we will use "human.encode.data" and "human.encode.data.descriptor" from C3 repository to make the compendium.
## These data sets are loaded automatically with the package.
## The gene expression data sets are in list format. Here every list element contains transcriptomic profile data of a cell or tissue type.
human.compendium<-make_gene_expression_compendium(gene.expression.data=human.encode.data, experiment.descriptor=human.encode.data.descriptor, expression.data.format="list")

## Finally we will make the marker gene compendium from the human.compendium
human.marker.gene.compendium<-make_marker_gene_compendium(compendium.data=human.compendium, specific.cutoff = 0.05, top.expressed.genes = 500)
summary(human.marker.gene.compendium)

}
