% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{point_options}
\alias{point_options}
\title{Point Options}
\usage{
point_options(c3, show = TRUE, r = 2.5, expand = TRUE, expand.r = 1.75,
  select.r = 4)
}
\arguments{
\item{show}{boolean}

\item{r}{numeric radius of point}

\item{expand}{boolean}

\item{expand.r}{numeric multiplier for radius expansion}

\item{select.r}{numeric multiplier for radius expansion}
}
\value{
c3
}
\description{
Modify point options
}

