% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{point_options}
\alias{point_options}
\title{Point Options}
\usage{
point_options(c3, show = TRUE, r = 2.5, expand = TRUE,
  expand.r = 1.75, select.r = 4)
}
\arguments{
\item{c3}{c3 htmlwidget object}

\item{show}{boolean}

\item{r}{numeric radius of point}

\item{expand}{boolean}

\item{expand.r}{numeric multiplier for radius expansion}

\item{select.r}{numeric multiplier for radius expansion}
}
\value{
c3
}
\description{
Modify point options
}
\examples{
data.frame(a = c(1,2,3,2), b = c(2,4,1,5)) \%>\%
  c3() \%>\%
  point_options(r = 5, expand.r = 2)

}
