% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{RColorBrewer.c3}
\alias{RColorBrewer.c3}
\title{RColorBrewer Palette}
\usage{
\method{RColorBrewer}{c3}(c3, pal = "Spectral")
}
\arguments{
\item{c3}{c3 htmlwidget object}

\item{pal}{character palette must match `RColorBrewer::brewer.pal.info`}
}
\value{
c3
}
\description{
Use RColorBrewer palettes
}
\examples{
\dontrun{
data.frame(a = c(1,2,3,2), b = c(2,4,1,5), c = c(5,3,4,1)) \%>\%
  c3() \%>\%
  RColorBrewer()
  }

}
\seealso{
Other c3: \code{\link{c3}}, \code{\link{grid.c3}},
  \code{\link{legend.c3}}, \code{\link{region.c3}},
  \code{\link{subchart.c3}}, \code{\link{tooltip.c3}},
  \code{\link{xAxis.c3}}, \code{\link{zoom.c3}}

Other RColorBrewer: \code{\link{RColorBrewer}}
}
