% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{RColorBrewer.c3}
\alias{RColorBrewer.c3}
\title{Title}
\usage{
\method{RColorBrewer}{c3}(c3, pal = "Spectral", ...)
}
\arguments{
\item{...}{}
}
\value{
c3
}
\description{
Title
}
\seealso{
Other RColorBrewer: \code{\link{RColorBrewer}}

Other c3: \code{\link{c3}}, \code{\link{grid.c3}},
  \code{\link{legend.c3}}, \code{\link{tooltip.c3}},
  \code{\link{xAxis.c3}}
}

