% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{RColorBrewer}
\alias{RColorBrewer}
\title{Modify grid and line elements on both x and y axis}
\usage{
RColorBrewer(x, ...)
}
\description{
This is an S3 method.
}
\seealso{
Other RColorBrewer: \code{\link{RColorBrewer.c3}}
}

