% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region.c3}
\alias{region.c3}
\title{Modify region elements on both x and y axis}
\usage{
\method{region}{c3}(c3, regions)
}
\arguments{
\item{regions}{data.frame with columns:
\itemize{
 \item{axis}{: character one of 'x', 'y', 'y2'}
 \item{start}{: numeric but must match defined axis type}
 \item{end}{: numeric but must match defined axis type}
 \item{class}{: character css class}
}}
}
\value{
c3
}
\description{
Regions are defined in multiple axis by passing a single `data.frame`
}
\examples{
\dontrun{
iris \%>\%
 c3(x='Sepal_Length', y='Sepal_Width', group = 'Species') \%>\%
 c3_scatter() \%>\%
 region(data.frame(axis = 'x',
                   start = 5,
                   end = 6))
}

}
\seealso{
Other c3: \code{\link{RColorBrewer.c3}}, \code{\link{c3}},
  \code{\link{grid.c3}}, \code{\link{legend.c3}},
  \code{\link{subchart.c3}}, \code{\link{tooltip.c3}},
  \code{\link{xAxis.c3}}, \code{\link{zoom.c3}}

Other region: \code{\link{region}}
}
