% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region}
\alias{region}
\alias{region.c3}
\title{Modify region elements on both x and y axis}
\usage{
region(c3, regions)

\method{region}{c3}(c3, regions)
}
\arguments{
\item{c3}{c3 htmlwidget object}

\item{regions}{data.frame with columns listed below. Any columns can be missing but results may be unexpected.
\itemize{
 \item{axis}{: character one of 'x', 'y', 'y2'}
 \item{start}{: numeric but must match defined axis type}
 \item{end}{: numeric but must match defined axis type}
 \item{class}{: character css class}
}}
}
\value{
c3
}
\description{
Regions are defined in multiple axis by passing a single `data.frame`
}
\examples{
iris \%>\%
 c3(x = 'Sepal_Length', y = 'Sepal_Width', group = 'Species') \%>\%
 c3_scatter() \%>\%
 region(data.frame(axis = 'x',
                   start = 5,
                   end = 6))

}
\seealso{
Other c3: \code{\link{RColorBrewer}}, \code{\link{c3}},
  \code{\link{grid}}, \code{\link{legend}},
  \code{\link{subchart}}, \code{\link{tooltip}},
  \code{\link{xAxis}}, \code{\link{zoom}}
}
\concept{c3}
\concept{region}
