% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subchart.R
\name{subchart.c3}
\alias{subchart.c3}
\title{Add Subchart}
\usage{
\method{subchart}{c3}(c3, height = 20, onbrush = NULL)
}
\arguments{
\item{height}{integer pixels}

\item{onbrush}{character js function, wrap character or character vector in JS()}
}
\value{
c3
}
\description{
subcharts are defined in multiple axis by passing a single `data.frame`
}
\examples{
\dontrun{
data.frame(a = abs(rnorm(20) * 10),
           b = abs(rnorm(20) * 10),
           date = seq(as.Date("2014-01-01"), by = "month", length.out = 20)) \%>\%
     c3(x = 'date') \%>\%
     subchart(height = 20)
}

}
\seealso{
Other c3: \code{\link{RColorBrewer.c3}}, \code{\link{c3}},
  \code{\link{grid.c3}}, \code{\link{legend.c3}},
  \code{\link{region.c3}}, \code{\link{tooltip.c3}},
  \code{\link{xAxis.c3}}, \code{\link{zoom.c3}}

Other subchart: \code{\link{subchart}}
}

