% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{tickAxis}
\alias{tickAxis}
\title{Title}
\usage{
tickAxis(c3, axis, ...)
}
\arguments{
\item{axis}{character 'x' or 'y' axis}

\item{...}{}

\item{format}{character js function, wrap character or character vector in JS()}
}
\value{
c3
}
\description{
Title
}

