% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip.c3}
\alias{tooltip.c3}
\title{Title}
\usage{
\method{tooltip}{c3}(c3, ...)
}
\arguments{
\item{show}{boolean}

\item{grouped}{boolean}

\item{format}{list with options:
\itemize{
 \item{title}{: character js function, wrap character or character vector in JS()}
 \item{name}{: character js function, wrap character or character vector in JS()}
 \item{value}{: character js function, wrap character or character vector in JS()}
}}

\item{position}{character js function, wrap character or character vector in JS()}

\item{contents}{character js function, wrap character or character vector in JS()}
}
\value{
c3
}
\description{
Title
}
\seealso{
Other c3: \code{\link{RColorBrewer.c3}}, \code{\link{c3}},
  \code{\link{grid.c3}}, \code{\link{legend.c3}},
  \code{\link{region.c3}}, \code{\link{xAxis.c3}}

Other tooltip: \code{\link{tooltip}}
}

