% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoom.c3}
\alias{zoom.c3}
\title{Add Zoom}
\usage{
\method{zoom}{c3}(c3, ...)
}
\arguments{
\item{enabled}{boolean}

\item{rescale}{boolean}

\item{extent}{numeric vector}

\item{onzoom}{character js function, wrap character or character vector in JS()}

\item{onzoomstart}{character js function, wrap character or character vector in JS()}

\item{onzoomend}{character js function, wrap character or character vector in JS()}
}
\value{
c3
}
\description{
Enable chart Zoom
}
\examples{
\dontrun{
data.frame(a = abs(rnorm(20) * 10),
           b = abs(rnorm(20) * 10),
           date = seq(as.Date("2014-01-01"), by = "month", length.out = 20)) \%>\%
     c3(x = 'date') \%>\%
     zoom()
}

}
\seealso{
Other c3: \code{\link{RColorBrewer.c3}}, \code{\link{c3}},
  \code{\link{grid.c3}}, \code{\link{legend.c3}},
  \code{\link{region.c3}}, \code{\link{subchart.c3}},
  \code{\link{tooltip.c3}}, \code{\link{xAxis.c3}}

Other zoom: \code{\link{zoom}}
}
